requirements.txt
opencv-python
mediapipe
pycaw
spotipy
python-dotenv